Convert 1.17 to Mojang Mappings #86
Also does some changes to how EntityPet constructor is found and the jumping field.

Jumping has to be handled differently because it will now be an obfuscated name.

EntityPet constructor lookup has to be modified because getNMSClass("World") will no longer be able to find the class due to package name changes I'm assuming.